+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-## TBD
+## [1.0.2] - 2018-10-23
+### Added
+* SshSettings has a new parameter that can be assigned: localhost.
+* Preconditions can check if an object is null or true with `checkNotNull` and `checkIsTrue`, respectively.
+
+## [1.0.1] - 2018-05-21
 ### Fixes
 * InvocationHandler now throws underlying exception. See [#2](https://github.com/HotelsDotCom/hcommon-ssh/issues/2). 
 
 
-## [1.0.0] 2018-04-24
+## [1.0.0] - 2018-04-24
 ### Added
 * First release.
 * Support for SSH tunnels.